Move only: Rework Integration Test and makes things more modular
Easier for writing future tests. There are no new tests added in this
commit. This only moves cpp tests into test_cpp, deletes read_file.rs
and moves common utilities to test_util.rs
