+++ integration_test/Cargo.toml
@@ -10,3 +10,4 @@ miniscript = {path = "../"}
 bitcoincore-rpc = {git = "https://github.com/sanket1729/rust-bitcoincore-rpc",rev = "bcc35944b3dd636cdff9710f90f8e0cfcab28f27"}
 bitcoin = "0.28.0-rc.1"
 log = "0.4"
+rand = "0.8.4"
